+++ R/messages.R
@@ -162,6 +162,13 @@ messages$valueNotPositive <- function(value, propertyName) {
   )
 }
 
+messages$yErrorValuesNegative <- function(n) {
+  cliFormat(
+    "{.val {n}} negative yErrorValue(s) found.",
+    "Negative values have been set to {.val NA}."
+  )
+}
+
 messages$molWeightErrorMessage <- function(quantityPath) {
   paste0("Unable to retrieve the molecular weight for: ", quantityPath)
 }